+++ include/new/ai_advanced.h
@@ -31,7 +31,7 @@ bool8 IsClassDoublesSpecific(u8 class);
 bool8 IsClassDoublesAttacker(u8 class);
 bool8 IsClassDamager(u8 class);
 u8 PredictBankFightingStyle(u8 bank);
-u8 PredictFightingStyle(const u16* const moves, const u8 ability, const u8 itemEffect, const u8 bank);
+u8 PredictFightingStyle(const u16* const moves, const u16 ability, const u8 itemEffect, const u8 bank);
 
 bool8 ShouldTrap(u8 bankAtk, u8 bankDef, u16 move, u8 class);
 bool8 ShouldRecover(u8 bankAtk, u8 bankDef, u16 move);
