+++ include/new/dexnav_data.h
@@ -403,7 +403,7 @@ struct DexnavHudData
     u16 species;
     u16 moveId[MAX_MON_MOVES];
     u16 heldItem;
-    u8 ability;
+    u16 ability;
     u8 potential;
     u8 searchLevel;
     u8 pokemonLevel;
