+++ src/accuracy_calc.c
@@ -21,7 +21,7 @@ extern const struct StatFractions gAccuracyStageRatios[];
 
 //This file's functions:
 static bool8 AccuracyCalcHelper(move_t, u8 bankDef);
-static u32 AccuracyCalcPassDefAbilityItemEffect(u16 move, u8 bankAtk, u8 bankDef, u8 defAbility, u8 defEffect);
+static u32 AccuracyCalcPassDefAbilityItemEffect(u16 move, u8 bankAtk, u8 bankDef, u16 defAbility, u8 defEffect);
 
 void atk01_accuracycheck(void)
 {
@@ -405,7 +405,7 @@ u32 AccuracyCalc(u16 move, u8 bankAtk, u8 bankDef)
 	return AccuracyCalcPassDefAbilityItemEffect(move, bankAtk, bankDef, ABILITY(bankDef), ITEM_EFFECT(bankDef));
 }
 
-static u32 AccuracyCalcPassDefAbilityItemEffect(u16 move, u8 bankAtk, u8 bankDef, u8 defAbility, u8 defEffect)
+static u32 AccuracyCalcPassDefAbilityItemEffect(u16 move, u8 bankAtk, u8 bankDef, u16 defAbility, u8 defEffect)
 {
 	u8 moveAcc;
 	s8 buff;
@@ -414,7 +414,7 @@ static u32 AccuracyCalcPassDefAbilityItemEffect(u16 move, u8 bankAtk, u8 bankDef
 	//u8 defEffect  = ITEM_EFFECT(bankDef);
 	u8 atkQuality = ITEM_QUALITY(bankAtk);
 	u8 defQuality = ITEM_QUALITY(bankDef);
-	u8 atkAbility = ABILITY(bankAtk);
+	u16 atkAbility = ABILITY(bankAtk);
 	//u8 defAbility = ABILITY(bankDef);
 	u8 moveSplit = CalcMoveSplit(gBankAttacker, move, bankDef);
 
@@ -537,7 +537,7 @@ static u32 AccuracyCalcPassDefAbilityItemEffect(u16 move, u8 bankAtk, u8 bankDef
 u32 VisualAccuracyCalc(u16 move, u8 bankAtk, u8 bankDef)
 {
 	u8 defEffect  = GetRecordedItemEffect(bankDef);
-	u8 defAbility = GetRecordedAbility(bankDef);
+	u16 defAbility = GetRecordedAbility(bankDef);
 	u32 acc = AccuracyCalcPassDefAbilityItemEffect(move, bankAtk, bankDef, defAbility, defEffect);
 
 	if (ABILITY(bankAtk) == ABILITY_NOGUARD || defAbility == ABILITY_NOGUARD
@@ -563,7 +563,7 @@ u32 VisualAccuracyCalc_NoTarget(u16 move, u8 bankAtk)
 	u32 calc;
 	u8 atkEffect  = ITEM_EFFECT(bankAtk);
 	u8 atkQuality = ITEM_QUALITY(bankAtk);
-	u8 atkAbility = ABILITY(bankAtk);
+	u16 atkAbility = ABILITY(bankAtk);
 	u8 moveSplit = SPLIT(move);
 
 	acc = STAT_STAGE(bankAtk, STAT_STAGE_ACC);
