+++ src/daycare.c
@@ -848,7 +848,7 @@ static u8 GetEggStepsToSubtract(void)
 	
 		if (species != SPECIES_NONE && species != SPECIES_EGG)
 		{
-			u8 ability = GetMonAbility(&gPlayerParty[i]);
+			u16 ability = GetMonAbility(&gPlayerParty[i]);
 
 			if (ability == ABILITY_MAGMAARMOR || ability == ABILITY_FLAMEBODY || ability == ABILITY_STEAMENGINE)
 			{
