+++ src/form_change.c
@@ -582,7 +582,7 @@ void HoldItemFormChange(struct Pokemon* mon, u16 item)
 	u16 targetSpecies = SPECIES_NONE;
 
 	u16 species = GetMonData(mon, MON_DATA_SPECIES, NULL);
-	u8 ability = GetMonAbility(mon);
+	u16 ability = GetMonAbility(mon);
 	u8 itemEffect = ItemId_GetHoldEffect(item);
 	u8 type = ItemId_GetHoldEffectParam(item);
